client,cli: decouple sustainability methods from common
