+++ aiven/client/client.py
@@ -3,14 +3,14 @@
 # This file is under the Apache License, Version 2.0.
 # See the file `LICENSE` for details.
 from __future__ import annotations
-from aiven.client.account_client import AivenAccountOrganisationClient
 
+from aiven.client.account_client import AivenAccountOrganisationClient
 from aiven.client.common_client import AivenCommonClient
 from aiven.client.opensearch_client import AivenOpenSearchClient
-from typing import TYPE_CHECKING
-
 from aiven.client.privatelink_client import AivenPrivateLinkClient
+from aiven.client.sustainability_client import AivenSustainabilityClient
 from aiven.client.user_client import AivenUserClient
+from typing import TYPE_CHECKING
 
 try:
     from .version import __version__
@@ -25,9 +25,10 @@ UNCHANGED = object()  # used as a sentinel value
 
 class AivenClient(
     AivenCommonClient,
-    AivenUserClient,
     AivenAccountOrganisationClient,
-    AivenPrivateLinkClient,
     AivenOpenSearchClient,
+    AivenPrivateLinkClient,
+    AivenSustainabilityClient,
+    AivenUserClient,
 ):
     pass
